[e2e] RHIDP-4264 - 1st PR: Cleanup and organization (#1721)
* cleanup

split rbac tests

+ api improvements

replace by beforeAll by RhdhApi

delete redundant toBeTruthy

moving constants at top of methods

fix RhdhApi api url

* fix exposure apis URLs

* rbac text to skip to fixme

* fix RhdhAuthHack

* fixing responseHelper

* rhdh api auth hack as singleton

* fix rhdh api

* minor improvements RHDH api

* -not rhdh api yet

back to responseHelper

* adapt RhdhAuthHack

retrieve removeMetadataFromResponse

delete useless awaits

---------
