+++ e2e-tests/playwright/e2e/plugins/rbac/rbac.spec.ts
@@ -6,280 +6,9 @@ import {
   RoleListPO,
   RoleOverviewPO,
 } from '../../../support/pageObjects/page-obj';
-import { Roles, Response } from '../../../support/pages/rbac';
+import { Roles } from '../../../support/pages/rbac';
 import { Common, setupBrowser } from '../../../utils/Common';
 import { UIhelper } from '../../../utils/UIhelper';
-
-test.describe.serial('Test RBAC plugin REST API', () => {
-  let common: Common;
-  let uiHelper: UIhelper;
-  let page: Page;
-  let responseHelper: Response;
-
-  test.beforeAll(async ({ browser, baseURL }, testInfo) => {
-    page = (await setupBrowser(browser, testInfo)).page;
-
-    uiHelper = new UIhelper(page);
-    common = new Common(page);
-
-    await common.loginAsGithubUser();
-
-    await uiHelper.openSidebar('Catalog');
-    const requestPromise = page.waitForRequest(
-      request =>
-        request.url() === `${baseURL}/api/search/query?term=` &&
-        request.method() === 'GET',
-    );
-    await uiHelper.openSidebar('Home');
-    const getRequest = await requestPromise;
-    const authToken = await getRequest.headerValue('Authorization');
-
-    responseHelper = new Response(authToken);
-  });
-
-  test('Test that roles and policies from GET request are what expected', async ({
-    request,
-  }) => {
-    const rolesResponse = await request.get(
-      '/api/permission/roles',
-      responseHelper.getSimpleRequest(),
-    );
-    const policiesResponse = await request.get(
-      '/api/permission/policies',
-      responseHelper.getSimpleRequest(),
-    );
-
-    await responseHelper.checkResponse(
-      await rolesResponse,
-      responseHelper.getExpectedRoles(),
-    );
-    await responseHelper.checkResponse(
-      await policiesResponse,
-      responseHelper.getExpectedPolicies(),
-    );
-  });
-
-  test('Create new role for rhdh-qe, change its name, and deny it from reading catalog entities', async ({
-    request,
-  }) => {
-    const members = ['user:default/rhdh-qe'];
-
-    const firstRole = {
-      memberReferences: members,
-      name: 'role:default/admin',
-    };
-    const rolePostResponse = await request.post(
-      '/api/permission/roles',
-      responseHelper.createRoleRequest(firstRole),
-    );
-    const newRole = {
-      memberReferences: members,
-      name: 'role:default/test',
-    };
-
-    const rolePutResponse = await request.put(
-      '/api/permission/roles/role/default/admin',
-      responseHelper.editRoleRequest(firstRole, newRole),
-    );
-
-    const newPolicy = {
-      entityReference: 'role:default/test',
-      permission: 'catalog-entity',
-      policy: 'read',
-      effect: 'deny',
-    };
-
-    const policyPostResponse = await request.post(
-      '/api/permission/policies',
-      responseHelper.createOrDeletePolicyRequest([newPolicy]),
-    );
-
-    expect(rolePostResponse.ok()).toBeTruthy();
-    expect(rolePutResponse.ok()).toBeTruthy();
-    expect(policyPostResponse.ok()).toBeTruthy();
-  });
-
-  test('Test catalog-entity read is denied', async () => {
-    await uiHelper.openSidebar('Catalog');
-    await uiHelper.selectMuiBox('Kind', 'Component');
-    await uiHelper.verifyTableIsEmpty();
-    await uiHelper.openSidebar('Create...');
-    await uiHelper.verifyText(
-      'No templates found that match your filter. Learn more about',
-      false,
-    );
-  });
-
-  test('Test catalog-entity creation is denied', async () => {
-    expect(
-      await uiHelper.isLinkVisible('Register Existing Component'),
-    ).toBeFalsy();
-  });
-
-  test('PUT catalog-entity read and POST create policies', async ({
-    request,
-  }) => {
-    const oldReadPolicy = [
-      { permission: 'catalog-entity', policy: 'read', effect: 'deny' },
-    ];
-    const newReadPolicy = [
-      { permission: 'catalog-entity', policy: 'read', effect: 'allow' },
-    ];
-    const readPutResponse = await request.put(
-      '/api/permission/policies/role/default/test',
-      responseHelper.editPolicyRequest(oldReadPolicy, newReadPolicy),
-    );
-
-    const createPolicy = [
-      {
-        entityReference: 'role:default/test',
-        permission: 'catalog.entity.create',
-        policy: 'create',
-        effect: 'allow',
-      },
-    ];
-    const createPostResponse = await request.post(
-      '/api/permission/policies',
-      responseHelper.createOrDeletePolicyRequest(createPolicy),
-    );
-
-    expect(readPutResponse.ok()).toBeTruthy();
-    expect(createPostResponse.ok()).toBeTruthy();
-  });
-
-  test('Test catalog-entity read is allowed', async () => {
-    await uiHelper.openSidebar('Catalog');
-    await uiHelper.selectMuiBox('Kind', 'API');
-    await uiHelper.clickLink('Nexus Repo Manager 3');
-  });
-
-  test('Test catalog-entity refresh is denied', async () => {
-    expect(
-      await uiHelper.isBtnVisibleByTitle('Schedule entity refresh'),
-    ).toBeFalsy();
-  });
-
-  test('Test catalog-entity create is allowed', async () => {
-    await uiHelper.openSidebar('Create...');
-    expect(
-      await uiHelper.isLinkVisible('Register Existing Component'),
-    ).toBeTruthy();
-  });
-
-  test('Test bad PUT and PUT catalog-entity update policy', async ({
-    request,
-  }) => {
-    const oldBadPolicy = [
-      { permission: 'catalog-entity', policy: 'refresh', effect: 'allow' },
-    ];
-    const newBadPolicy = [
-      { permission: 'catalog-entity', policy: 'read', effect: 'allow' },
-    ];
-    const badPutResponse = await request.put(
-      '/api/permission/policies/role/default/test',
-      responseHelper.editPolicyRequest(oldBadPolicy, newBadPolicy),
-    );
-
-    const oldGoodPolicy = [
-      {
-        permission: 'catalog.entity.create',
-        policy: 'create',
-        effect: 'allow',
-      },
-    ];
-    const newGoodPolicy = [
-      {
-        permission: 'catalog.entity.refresh',
-        policy: 'update',
-        effect: 'allow',
-      },
-    ];
-    const goodPutResponse = await request.put(
-      '/api/permission/policies/role/default/test',
-      responseHelper.editPolicyRequest(oldGoodPolicy, newGoodPolicy),
-    );
-
-    expect(badPutResponse.ok()).toBeFalsy();
-    expect(goodPutResponse.ok()).toBeTruthy();
-  });
-
-  test('Test that the bad PUT didnt go through and catalog-entities can be read', async () => {
-    await uiHelper.openSidebar('Home');
-    await uiHelper.openSidebar('Create...');
-    expect(
-      await uiHelper.isTextVisible(
-        'No templates found that match your filter. Learn more about',
-      ),
-    ).toBeFalsy();
-  });
-
-  test('Test that the good PUT request went through and catalog-entities can be refreshed', async () => {
-    await uiHelper.openSidebar('Catalog');
-    await uiHelper.selectMuiBox('Kind', 'API');
-    await uiHelper.clickLink('Nexus Repo Manager 3');
-    expect(
-      await uiHelper.isBtnVisibleByTitle('Schedule entity refresh'),
-    ).toBeTruthy();
-  });
-
-  test('Test that the good PUT request went through and catalog-entities cant be created', async () => {
-    await uiHelper.openSidebar('Create...');
-    expect(
-      await uiHelper.isLinkVisible('Register Existing Component'),
-    ).toBeFalsy();
-  });
-
-  test('DELETE catalog-entity update policy', async ({ request }) => {
-    const deletePolicies = [
-      {
-        entityReference: 'role:default/test',
-        permission: 'catalog.entity.refresh',
-        policy: 'update',
-        effect: 'allow',
-      },
-    ];
-    const deleteResponse = await request.delete(
-      '/api/permission/policies/role/default/test',
-      responseHelper.createOrDeletePolicyRequest(deletePolicies),
-    );
-
-    expect(deleteResponse.ok()).toBeTruthy();
-  });
-
-  test('Test catalog-entity refresh is denied after DELETE', async () => {
-    await uiHelper.openSidebar('Catalog');
-    await uiHelper.selectMuiBox('Kind', 'API');
-    await uiHelper.clickLink('Nexus Repo Manager 3');
-    expect(await uiHelper.isBtnVisible('Schedule entity refresh')).toBeFalsy();
-  });
-
-  test.afterAll(
-    'Cleanup by deleting all new policies and roles',
-    async ({ request }) => {
-      const remainingPoliciesResponse = await request.get(
-        '/api/permission/policies/role/default/test',
-        responseHelper.getSimpleRequest(),
-      );
-
-      const remainingPolicies = await responseHelper.removeMetadataFromResponse(
-        remainingPoliciesResponse,
-      );
-
-      const deleteRemainingPolicies = await request.delete(
-        '/api/permission/policies/role/default/test',
-        responseHelper.createOrDeletePolicyRequest(remainingPolicies),
-      );
-
-      const deleteRole = await request.delete(
-        '/api/permission/roles/role/default/test',
-        responseHelper.getSimpleRequest(),
-      );
-
-      expect(deleteRemainingPolicies.ok()).toBeTruthy();
-      expect(deleteRole.ok()).toBeTruthy();
-    },
-  );
-});
 
 test.describe
   .serial('Test RBAC plugin: load permission policies and conditions from files', () => {
@@ -371,7 +100,7 @@ test.describe
 
     await expect(page.locator('header')).toContainText('user:rhdh-qe-2');
     await page.getByTestId('menu-button').click();
-    const unregisterUserOwned = await page.getByText('Unregister entity');
+    const unregisterUserOwned = page.getByText('Unregister entity');
     await expect(unregisterUserOwned).toBeEnabled();
 
     await page.getByText('Unregister entity').click();
@@ -386,7 +115,7 @@ test.describe
       'janus-qe/rhdh-qe-2-team',
     );
     await page.getByTestId('menu-button').click();
-    const unregisterGroupOwned = await page.getByText('Unregister entity');
+    const unregisterGroupOwned = page.getByText('Unregister entity');
     await expect(unregisterGroupOwned).toBeDisabled();
   });
 
@@ -523,92 +252,104 @@ test.describe.serial('Test RBAC plugin as an admin user', () => {
   });
 
   //FIXME
-  test.skip('Admin cannot create a role if there are no rules defined for the selected resource type.', async () => {
-    await uiHelper.clickButton('Create');
-    await uiHelper.verifyHeading('Create role');
+  test.fixme(
+    'Admin cannot create a role if there are no rules defined for the selected resource type.',
+    async () => {
+      await uiHelper.clickButton('Create');
+      await uiHelper.verifyHeading('Create role');
 
-    await page.fill(RoleFormPO.roleName, 'test-role');
-    await uiHelper.clickButton('Next');
-    await page.fill(RoleFormPO.addUsersAndGroups, 'guest user');
-    await page.click(RoleFormPO.selectMember('Guest User'));
-    await uiHelper.clickButton('Next');
+      await page.fill(RoleFormPO.roleName, 'test-role');
+      await uiHelper.clickButton('Next');
+      await page.fill(RoleFormPO.addUsersAndGroups, 'guest user');
+      await page.click(RoleFormPO.selectMember('Guest User'));
+      await uiHelper.clickButton('Next');
 
-    await page.click(RoleFormPO.selectPermissionPolicyPlugin(0), {
-      timeout: 100000,
-    });
-    await uiHelper.optionSelector('catalog');
+      await page.click(RoleFormPO.selectPermissionPolicyPlugin(0), {
+        timeout: 100000,
+      });
+      await uiHelper.optionSelector('catalog');
 
-    await page.click(RoleFormPO.selectPermissionPolicyPermission(0), {
-      timeout: 100000,
-    });
-    await uiHelper.optionSelector('catalog.entity.create');
+      await page.click(RoleFormPO.selectPermissionPolicyPermission(0), {
+        timeout: 100000,
+      });
+      await uiHelper.optionSelector('catalog.entity.create');
 
-    await expect(page.getByLabel('configure-access')).not.toBeVisible();
-    await uiHelper.clickButton('Cancel');
-  });
+      await expect(page.getByLabel('configure-access')).not.toBeVisible();
+      await uiHelper.clickButton('Cancel');
+    },
+  );
 
-  test.skip('As an RHDH admin, I want to be able to restrict access by using the Not condition to part of the plugin, so that some information is protected from unauthorized access.', async () => {
-    await rolesHelper.createRoleWithNotPermissionPolicy('test-role');
-    await page.locator(HomePagePO.searchBar).waitFor({ state: 'visible' });
-    await page.locator(HomePagePO.searchBar).fill('test-role');
-    await uiHelper.verifyHeading('All roles (1)');
+  test.fixme(
+    'As an RHDH admin, I want to be able to restrict access by using the Not condition to part of the plugin, so that some information is protected from unauthorized access.',
+    async () => {
+      await rolesHelper.createRoleWithNotPermissionPolicy('test-role');
+      await page.locator(HomePagePO.searchBar).waitFor({ state: 'visible' });
+      await page.locator(HomePagePO.searchBar).fill('test-role');
+      await uiHelper.verifyHeading('All roles (1)');
 
-    await rolesHelper.deleteRole('role:default/test-role');
-  });
+      await rolesHelper.deleteRole('role:default/test-role');
+    },
+  );
 
-  test.skip('As an RHDH admin, I want to be able to edit the access rule, so I can keep it up to date and be able to add more plugins in the future.', async () => {
-    await rolesHelper.createRoleWithNotPermissionPolicy('test-role');
-    await page.locator(HomePagePO.searchBar).waitFor({ state: 'visible' });
-    await page.locator(HomePagePO.searchBar).fill('test-role');
-    await uiHelper.verifyHeading('All roles (1)');
+  test.fixme(
+    'As an RHDH admin, I want to be able to edit the access rule, so I can keep it up to date and be able to add more plugins in the future.',
+    async () => {
+      await rolesHelper.createRoleWithNotPermissionPolicy('test-role');
+      await page.locator(HomePagePO.searchBar).waitFor({ state: 'visible' });
+      await page.locator(HomePagePO.searchBar).fill('test-role');
+      await uiHelper.verifyHeading('All roles (1)');
 
-    await page.click(RoleListPO.editRole('role:default/test-role'));
-    await uiHelper.verifyHeading('Edit Role');
-    await uiHelper.clickButton('Next');
-    await uiHelper.clickButton('Next');
+      await page.click(RoleListPO.editRole('role:default/test-role'));
+      await uiHelper.verifyHeading('Edit Role');
+      await uiHelper.clickButton('Next');
+      await uiHelper.clickButton('Next');
 
-    await page.getByLabel('configure-access').click();
-    await page.getByRole('button', { name: 'Condition' }).click();
-    await page.getByTestId('rules-sidebar').getByLabel('Open').click();
-    await page.getByText('HAS_SPEC').click();
-    await page.getByLabel('key *').click();
-    await page.getByLabel('key *').fill('lifecycle');
-    await page.getByTestId('save-conditions').click();
+      await page.getByLabel('configure-access').click();
+      await page.getByRole('button', { name: 'Condition' }).click();
+      await page.getByTestId('rules-sidebar').getByLabel('Open').click();
+      await page.getByText('HAS_SPEC').click();
+      await page.getByLabel('key *').click();
+      await page.getByLabel('key *').fill('lifecycle');
+      await page.getByTestId('save-conditions').click();
 
-    await uiHelper.clickButton('Next');
-    await uiHelper.clickButton('Save');
-    await uiHelper.verifyText(
-      'Role role:default/test-role updated successfully',
-    );
+      await uiHelper.clickButton('Next');
+      await uiHelper.clickButton('Save');
+      await uiHelper.verifyText(
+        'Role role:default/test-role updated successfully',
+      );
 
-    await rolesHelper.deleteRole('role:default/test-role');
-  });
+      await rolesHelper.deleteRole('role:default/test-role');
+    },
+  );
 
-  test.skip('As an RHDH admin, I want to be able to remove an access rule from an existing permission policy.', async () => {
-    await rolesHelper.createRoleWithPermissionPolicy('test-role');
-    await page.locator(HomePagePO.searchBar).waitFor({ state: 'visible' });
-    await page.locator(HomePagePO.searchBar).fill('test-role');
-    await uiHelper.verifyHeading('All roles (1)');
+  test.fixme(
+    'As an RHDH admin, I want to be able to remove an access rule from an existing permission policy.',
+    async () => {
+      await rolesHelper.createRoleWithPermissionPolicy('test-role');
+      await page.locator(HomePagePO.searchBar).waitFor({ state: 'visible' });
+      await page.locator(HomePagePO.searchBar).fill('test-role');
+      await uiHelper.verifyHeading('All roles (1)');
 
-    await page.click(RoleListPO.editRole('role:default/test-role'));
-    await uiHelper.verifyHeading('Edit Role');
-    await uiHelper.clickButton('Next');
-    await uiHelper.clickButton('Next');
+      await page.click(RoleListPO.editRole('role:default/test-role'));
+      await uiHelper.verifyHeading('Edit Role');
+      await uiHelper.clickButton('Next');
+      await uiHelper.clickButton('Next');
 
-    await page.getByLabel('configure-access').click();
-    await page.getByRole('button', { name: 'Remove' }).nth(2).click();
-    await page.getByTestId('save-conditions').click();
-    await uiHelper.verifyText('Configure access (2 rules)');
+      await page.getByLabel('configure-access').click();
+      await page.getByRole('button', { name: 'Remove' }).nth(2).click();
+      await page.getByTestId('save-conditions').click();
+      await uiHelper.verifyText('Configure access (2 rules)');
 
-    await uiHelper.clickButton('Next');
-    await uiHelper.clickButton('Save');
+      await uiHelper.clickButton('Next');
+      await uiHelper.clickButton('Save');
 
-    await uiHelper.verifyText(
-      'Role role:default/test-role updated successfully',
-    );
+      await uiHelper.verifyText(
+        'Role role:default/test-role updated successfully',
+      );
 
-    await rolesHelper.deleteRole('role:default/test-role');
-  });
+      await rolesHelper.deleteRole('role:default/test-role');
+    },
+  );
 
   test.afterAll(async () => {
     await page.close();
